ui: move window into ui
