+++ ui/inputs.go
@@ -1,4 +1,4 @@
-package window
+package ui
 
 func (win *Window) HandleInputs() {
